compiler: Lower `nowrite` attribute as `inaccessiblememonly`
The TBAA-based approach, with it marking the memory as immutable,
somehow led to the loop in Core.wait_until_mu() being optimised
out with the extended pass pipeline, making the
CoredeviceTest.test_wait_for_rtio_counter test case fail.
Presumably, the entire loop got eliminated, as infinite loops
would be UB, although I haven't verified this.

It's not entirely clear to me whether this is a bug in LLVM 6,
or an intended feature of being marked only with an "immutable"
TBAA node, as !tbaa metadata on call instructions doesn't even
seem to be documented (neither in LangRef nor AliasAnalysis).

`nowrite` is a bit of a misnomer now and should be renamed, but
I didn't want to break compatibility with people's out-of-tree
drivers just yet.