gui: Support dashboard applets on macOS (previously crashed)
Embedding of foreign windows does not seem to be supported at all
on macOS. The previous implementation on macOS had the effect of
dereferencing a pointer in the address space of the applet process
in the dashboard process, crashing as a result.

This commit changes applets just to be launched as stand-alone
subprocesses on macOS instead, with some integration to link their
life cycle to the checkboxes in the applet list.